# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Social Distribution - Ctrl + Alt + Defeat API",
      "description": "API for our backend"
    },
    "paths": {
      "/api/authors/": {
        "get": {
          "summary": "Get all authors (REMOTE)",
          "parameters": [
            {
              "name": "page",
              "in": "query",
              "type": "integer",
              "description": "Page number",
              "required": false
            },
            {
              "name": "size",
              "in": "query",
              "type": "integer",
              "description": "Number of items per page",
              "required": false
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response",
              "schema": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "authors"
                  },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "author"
                        },
                        "id": {
                          "type": "string",
                          "example": "http://127.0.0.1:5454/authors/1d698d25ff008f7538453c120f581471"
                        },
                        "url": {
                          "type": "string",
                          "example": "http://127.0.0.1:5454/authors/1d698d25ff008f7538453c120f581471"
                        },
                        "host": {
                          "type": "string",
                          "example": "http://127.0.0.1:5454/"
                        },
                        "displayName": {
                          "type": "string",
                          "example": "Greg Johnson"
                        },
                        "github": {
                          "type": "string",
                          "example": "http://github.com/gjohnson"
                        },
                        "profileImage": {
                          "type": "string",
                          "example": "https://i.imgur.com/k7XVwpB.jpeg"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "tags": ["Author"]
        }
      },
...
          "/api/posts/restricted": {
            "get": {
              "summary": "Get restricted users for a post",
              "parameters": [
                {
                  "name": "post_id",
                  "in": "query",
                  "type": "string",
                  "description": "ID of the post",
                  "required": true
                }
              ],
              "responses": {
                "200": {
                  "description": "Successful response",
                  "schema": {
                    "type": "object"
                  }
                }
              },
              "tags": ["Posts"]
            }
          },
...
          "/api/posts/unrestrict/{post_id}/{username}": {
            "delete": {
              "summary": "Unrestrict user from a post",
              "parameters": [
                {
                  "name": "post_id",
                  "in": "path",
                  "type": "string",
                  "description": "ID of the post",
                  "required": true
                },
                {
                  "name": "username",
                  "in": "path",
                  "type": "string",
                  "description": "Username of the user to unrestrict",
                  "required": true
                }
              ],
              "responses": {
                "200": {
                  "description": "Successful response",
                  "schema": {
                    "type": "object"
                  }
                }
              },
              "tags": ["Posts"]
            }
          },
...
          "/api/posts/manage": {
            "get": {
              "summary": "Get posts managed by the user",
              "parameters": [
                {
                  "name": "author_id",
                  "in": "query",
                  "type": "string",
                  "description": "ID of the user",
                  "required": true
                }
              ],
              "responses": {
                "200": {
                  "description": "Successful response",
                  "schema": {
                    "type": "object"
                  }
                }
              },
              "tags": ["Posts"]
            }
          },
          "/api/posts": {
            "get": {
              "summary": "Get posts",
              "parameters": [
                {
                  "name": "author_id",
                  "in": "path",
                  "type": "string",
                  "description": "ID of the user",
                  "required": true
                }
              ],
              "responses": {
                "200": {
                  "description": "Successful response",
                  "schema": {
                    "type": "object"
                  }
                }
              },
              "tags": ["Posts"]
            },
            "post": {
              "summary": "Create a new post",
              "responses": {
                "200": {
                  "description": "Successful response",
                  "schema": {
                    "type": "object"
                  }
                }
              },
              "tags": ["Posts"]
            }
          },
...
          "/api/authors/{author_id}/posts/{post_id}/image": {
            "get": {
              "summary": "Get image for a post (REMOTE)",
              "parameters": [
                {
                  "name": "author_id",
                  "in": "path",
                  "type": "string",
                  "description": "Author ID",
                  "required": true
                },
                {
                  "name": "post_id",
                  "in": "path",
                  "type": "string",
                  "description": "Post ID",
                  "required": true
                }
              ],
              "responses": {
                "200": {
                  "description": "Successful response",
                  "schema": {
                    "type": "object"
                  }
                }
              },
              "tags": ["Posts"]
            }
          },
...
          "/api/authors/{author_id}/posts/{post_id}": {
            "get": {
              "summary": "Get an individual post (REMOTE)",
              "parameters": [
                {
                  "name": "author_id",
                  "in": "path",
                  "type": "string",
                  "description": "Author ID",
                  "required": true
                },
                {
                  "name": "post_id",
                  "in": "path",
                  "type": "string",
                  "description": "Post ID",
                  "required": true
                }
              ],
              "responses": {
                "200": {
                  "description": "Successful response",
                  "schema": {
                    "type": "object"
                  }
                }
              },
              "tags": ["Posts"]
            }
          },
...
    }
  }